requirements.txt
aiogram==3.4
gspread
google-auth
python-dotenv
pillow
